+++ packages/arcgis-rest-request/src/request.ts
@@ -4,75 +4,10 @@
 import { encodeFormData } from "./utils/encode-form-data";
 import { encodeQueryString } from "./utils/encode-query-string";
 import { requiresFormData } from "./utils/process-params";
+import { checkForErrors } from "./utils/check-for-errors";
 import { ArcGISRequestError } from "./utils/ArcGISRequestError";
-import { IRetryAuthError } from "./utils/retryAuthError";
-import { HTTPMethods, IParams, ITokenRequestOptions } from "./utils/params";
-
-/**
- * Authentication can be supplied to `request` via [`UserSession`](../../auth/UserSession/) or [`ApplicationSession`](../../auth/ApplicationSession/). Both classes extend `IAuthenticationManager`.
- * ```js
- * const session = new UserSession({
- *   username: "jsmith",
- *   password: "123456",
- *   // optional
- *   portal: "https://[yourserver]/arcgis/sharing/rest"
- * })
- *
- * request(url, { authentication: session })
- * ```
- */
-export interface IAuthenticationManager {
-  /**
-   * Defaults to 'https://www.arcgis.com/sharing/rest'.
-   */
-  portal: string;
-  getToken(url: string, requestOptions?: ITokenRequestOptions): Promise<string>;
-}
-
-/**
- * Options for the `request()` method.
- */
-export interface IRequestOptions {
-  /**
-   * Additional parameters to pass in the request.
-   */
-  params?: IParams;
-
-  /**
-   * The HTTP method to send the request with.
-   */
-  httpMethod?: HTTPMethods;
-
-  /**
-   * Return the raw [response](https://developer.mozilla.org/en-US/docs/Web/API/Response)
-   */
-  rawResponse?: boolean;
-
-  /**
-   * The instance of `IAuthenticationManager` to use to authenticate this request.
-   */
-  authentication?: IAuthenticationManager;
-
-  /**
-   * Base url for the portal you want to make the request to. Defaults to 'https://www.arcgis.com/sharing/rest'.
-   */
-  portal?: string;
-
-  /**
-   * The implementation of `fetch` to use. Defaults to a global `fetch`.
-   */
-  fetch?: (input: RequestInfo, init?: RequestInit) => Promise<Response>;
-
-  /**
-   * If the length of a GET request's URL exceeds `maxUrlLength` the request will use POST instead.
-   */
-  maxUrlLength?: number;
-
-  /**
-   * Additional [Headers](https://developer.mozilla.org/en-US/docs/Web/API/Headers) to pass into the request.
-   */
-  headers?: { [key: string]: any };
-}
+import { IRequestOptions } from "./utils/IRequestOptions";
+import { IParams } from "./utils/IParams";
 
 export const NODEJS_DEFAULT_REFERER_HEADER = `@esri/arcgis-rest-js`;
 
@@ -98,13 +33,20 @@ export const NODEJS_DEFAULT_REFERER_HEADER = `@esri/arcgis-rest-js`;
  * @param requestOptions - Options for the request, including parameters relevant to the endpoint.
  * @returns A Promise that will resolve with the data from the response.
  */
+export function request(url: string | IRequestOptions): Promise<any>;
 export function request(
-  url: string,
-  requestOptions: IRequestOptions = { params: { f: "json" } }
+  requestUrl: string,
+  requestOptions: IRequestOptions
+): Promise<any>;
+export function request(
+  requestUrl: string | IRequestOptions,
+  requestOptions?: IRequestOptions
 ): Promise<any> {
+  let url = typeof requestUrl === "string" ? requestUrl : requestOptions.url;
   const options: IRequestOptions = {
     httpMethod: "POST",
-    ...requestOptions
+    ...{ params: { f: "json" } },
+    ...(typeof url === "string" ? requestOptions : url)
   };
 
   const missingGlobals: string[] = [];
@@ -254,117 +196,4 @@ export function request(
         return data;
       }
     });
-}
-
-export class ArcGISAuthError extends ArcGISRequestError {
-  /**
-   * Create a new `ArcGISAuthError`  object.
-   *
-   * @param message - The error message from the API
-   * @param code - The error code from the API
-   * @param response - The original response from the API that caused the error
-   * @param url - The original url of the request
-   * @param options - The original options of the request
-   */
-  constructor(
-    message = "AUTHENTICATION_ERROR",
-    code: string | number = "AUTHENTICATION_ERROR_CODE",
-    response?: any,
-    url?: string,
-    options?: IRequestOptions
-  ) {
-    super(message, code, response, url, options);
-    this.name = "ArcGISAuthError";
-    this.message =
-      code === "AUTHENTICATION_ERROR_CODE" ? message : `${code}: ${message}`;
-  }
-
-  public retry(getSession: IRetryAuthError, retryLimit = 3) {
-    let tries = 0;
-
-    const retryRequest = (resolve: any, reject: any) => {
-      getSession(this.url, this.options)
-        .then(session => {
-          const newOptions = {
-            ...this.options,
-            ...{ authentication: session }
-          };
-
-          tries = tries + 1;
-          return request(this.url, newOptions);
-        })
-        .then(response => {
-          resolve(response);
-        })
-        .catch(e => {
-          if (e.name === "ArcGISAuthError" && tries < retryLimit) {
-            retryRequest(resolve, reject);
-          } else if (e.name === "ArcGISAuthError" && tries >= retryLimit) {
-            reject(this);
-          } else {
-            reject(e);
-          }
-        });
-    };
-
-    return new Promise((resolve, reject) => {
-      retryRequest(resolve, reject);
-    });
-  }
-}
-
-/**
- * Checks for errors in a JSON response from the ArcGIS REST API. If there are no errors, it will return the `data` passed in. If there is an error, it will throw an `ArcGISRequestError` or `ArcGISAuthError`.
- *
- * @param data The response JSON to check for errors.
- * @param url The url of the original request
- * @param params The parameters of the original request
- * @param options The options of the original request
- * @returns The data that was passed in the `data` parameter
- */
-export function checkForErrors(
-  response: any,
-  url?: string,
-  params?: IParams,
-  options?: IRequestOptions
-): any {
-  // this is an error message from billing.arcgis.com backend
-  if (response.code >= 400) {
-    const { message, code } = response;
-    throw new ArcGISRequestError(message, code, response, url, options);
-  }
-
-  // error from ArcGIS Online or an ArcGIS Portal or server instance.
-  if (response.error) {
-    const { message, code, messageCode } = response.error;
-    const errorCode = messageCode || code || "UNKNOWN_ERROR_CODE";
-
-    if (
-      code === 498 ||
-      code === 499 ||
-      messageCode === "GWM_0003" ||
-      (code === 400 && message === "Unable to generate token.")
-    ) {
-      throw new ArcGISAuthError(message, errorCode, response, url, options);
-    }
-
-    throw new ArcGISRequestError(message, errorCode, response, url, options);
-  }
-
-  // error from a status check
-  if (response.status === "failed" || response.status === "failure") {
-    let message: string;
-    let code: string = "UNKNOWN_ERROR_CODE";
-
-    try {
-      message = JSON.parse(response.statusMessage).message;
-      code = JSON.parse(response.statusMessage).code;
-    } catch (e) {
-      message = response.statusMessage || response.message;
-    }
-
-    throw new ArcGISRequestError(message, code, response, url, options);
-  }
-
-  return response;
 }
