+++ packages/arcgis-rest-request/src/utils/GrantTypes.ts
@@ -1,0 +1,5 @@
+export type GrantTypes =
+  | "authorization_code"
+  | "refresh_token"
+  | "client_credentials"
+  | "exchange_refresh_token";
