+++ packages/arcgis-rest-request/src/utils/IFetchTokenParams.ts
@@ -1,0 +1,11 @@
+import { GrantTypes } from "./GrantTypes";
+import { IParams } from "./IParams";
+
+export interface IFetchTokenParams extends IParams {
+  client_id: string;
+  client_secret?: string;
+  grant_type: GrantTypes;
+  redirect_uri?: string;
+  refresh_token?: string;
+  code?: string;
+}
