+++ packages/arcgis-rest-request/src/utils/IGenerateTokenParams.ts
@@ -1,0 +1,9 @@
+import { IParams } from "./IParams";
+
+export interface IGenerateTokenParams extends IParams {
+  username?: string;
+  password?: string;
+  expiration?: number;
+  token?: string;
+  serverUrl?: string;
+}
