+++ packages/arcgis-rest-request/src/utils/IParams.ts
@@ -1,0 +1,6 @@
+import { ResponseFormats } from "./ResponseFormats";
+
+export interface IParams {
+  f?: ResponseFormats;
+  [key: string]: any;
+}
