+++ packages/arcgis-rest-request/src/utils/ResponseFormats.ts
@@ -1,0 +1,10 @@
+/**
+ * Valid response formats for the `f` parameter.
+ */
+export type ResponseFormats =
+  | "json"
+  | "geojson"
+  | "text"
+  | "html"
+  | "image"
+  | "zip";
