+++ packages/arcgis-rest-request/src/utils/get-portal-url.ts
@@ -1,6 +1,6 @@
 /* Copyright (c) 2017 Environmental Systems Research Institute, Inc.
  * Apache-2.0 */
-import { IRequestOptions } from "../request";
+import { IRequestOptions } from "./IRequestOptions";
 import { cleanUrl } from "./clean-url";
 /**
  * Helper that returns the appropriate portal url for a given request. `requestOptions.portal` is given
