+++ packages/arcgis-rest-request/src/utils/with-request-options.ts
@@ -1,0 +1,3 @@
+export function withOptions() {
+  // test
+}
